Add TCPConnectionPool to avro.Client.
Existing socket abstractions in pycernan.avro do not provide any level
of isolation for callers.  In practice this translates to a single
timeout, for example, resulting in subsequent publishers receiving acks
out of order when a Client is reused without first being reconstructed.

Rather than forcing users to think about this and to make it easier for
users to take advantage of connection level threading within Cernan, we
introduce a connection pool to avro.Client which has the following properties:

* Fork-safe as long as the Client isn't used prior forking.

* Established connections are preferred over establishing new
connections.

* Exceptions (on connect or during publication) are re-raised back to
the caller.

* Whenever a pool connection experiences an Exception, either operational
or specific to Cernan's wire protocol, the connection will be
re-established in a lazy fashion.  That is, on the next operation
assigned to that socket.